tourism_website/scripts/requirements.txt
firebase-admin==6.5.0
google-auth==2.34.0
requests==2.32.3
